Refactor configuration context and cleanup naming convention
