+++ core/App/index.ts
@@ -12,21 +12,22 @@ export { testIdWithKey } from './utils/testable'
 export { default as toastConfig } from './components/toast/ToastConfig'
 export { initLanguages, initStoredLanguage, defaultTranslationResources } from './localization'
 import RootStack from './navigators/RootStack'
+import OnboardingPages from './screens/OnboardingPages'
+import Splash from './screens/Splash'
+import Terms from './screens/Terms'
 export { ColorPallet } from './theme'
 export { default as indyLedgers } from '../configs/ledgers/indy'
 export { default as ErrorModal } from './components/modals/ErrorModal'
 export { StoreProvider, StoreContext } from './contexts/store'
 export type { StoreProviderProps } from './contexts/store'
-
 export { defaultTheme } from './theme'
 export type { Theme } from './theme'
 export { ThemeProvider, useThemeContext } from './utils/themeContext'
-export { ConfigurationProvider, useConfigurationContext } from './utils/configurationContext'
-export type { ConfigurationContext } from './utils/configurationContext'
+export { ConfigurationProvider, useConfiguration } from './contexts/configuration'
+export type { ConfigurationContext } from './contexts/configuration'
 export { default as defaultTerms } from './screens/Terms'
-export { default as defaultSplashScreen } from './screens/Splash'
 
-export { pages as defaultOnboardingPages, createStyle as createOnboardingPagesStye } from './screens/OnboardingPages'
+export { createStyles } from './screens/OnboardingPages'
 
 export { LocalStorageKeys } from './constants'
 export type { GenericFn } from './types/fn'
@@ -35,6 +36,6 @@ export type { Onboarding as OnboardingState } from './types/state'
 export { Screens } from './types/navigators'
 export type { AuthenticateStackParams } from './types/navigators'
 
-export { InfoTextBox, CheckBoxRow, Button, ButtonType, RootStack, Agent, AgentProvider }
+export { InfoTextBox, CheckBoxRow, Button, ButtonType, RootStack, Agent, AgentProvider, OnboardingPages, Splash, Terms }
 
 export type { OnboardingStyleSheet }
